feat: 서버 부팅시 비어있는 stockHistory 채우는 함수 추가
