+++ src/stock/stock.module.ts
@@ -4,10 +4,11 @@ import { StockService } from './stock.service';
 import { StockController } from './stock.controller';
 import { ChartModule } from './chart/chart.module';
 import { StockLimitService } from './order/services/stock-limit.service';
+import { StockHistorySchedulerService } from './scheduler/stock-history.scheduler.service';
 
 @Module({
     imports: [OrderModule, ChartModule],
     controllers: [StockController],
-    providers: [StockService, StockLimitService],
+    providers: [StockService, StockLimitService, StockHistorySchedulerService],
 })
 export class StockModule {}
